Implement download sample files for v2
